requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
gymnasium>=0.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tensorboard>=2.13.0
pytest>=7.3.0
tqdm>=4.65.0
networkx>=3.1
pandas>=2.0.0
scikit-learn>=1.2.0
